+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2023-10-28
+
+### Fixed
+
+* Monument structure description
+* Edit Structure Rules macro
+* Structure item availability rules now limit to maximum buildings occurrences, not maximum modifiers
+
+### Added
+
+* Vance & Kerenshara Structures
+
 ## [0.10.2] - 2023-10-28
 
 ### Fixed